Changed the way we handle stripe customer ids.
The previous method required that we remove the customer and add them
every time we want to change their subscription.
This method loses significant data of how users subscribe. This also
broke stripes pro-rate features when switching between plans in a
subscription.

We now always ensure that we have single customer id for any user that
has ever signed up.