added symbolBuilder with authoritative svg symbol collection
As part of the effort to de-crapify cardData, symbolBuilder has been
added to break out that component.

The cardSymbols directory is the complete collection of svg symbols. It
is being checked into source control as collecting and maintaining it is
painful enough to warrant the repo bloat.

Notably, the first node package on the backend. As a piece of the
production pipeline rather than the api, we get the convenience of node
without the headache of making it fast or sane.