faster latest extrema queries
Previously, extrema would check the entire selection of prices with an
expensive query, this filters it down to the last week to dramatically
lower the amount of work done.

This does introduce a breaking change in that price data needs to have
been acquired within the last week or the query will fail.